# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Testhub API
  version: 0.6.0
  description: |
    Core API for Testhub (runs, results ingestion).

    Authentication:
    - Protected endpoints require an API key via the `x-api-key` header.
    - Authorization is organization-scoped; non-owned resources return 404.
  license:
    name: MIT
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      operationId: listProjects
      summary: List projects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Projects]
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects/{projectId}:
    get:
      tags: [Projects]
      operationId: getProject
      summary: Get project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Projects]
      operationId: updateProject
      summary: Update project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Projects]
      operationId: deleteProject
      summary: Delete a project
      description: |
        Deletes a project and all associated data (runs, test cases, results).
        This operation cannot be undone.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: No Content - Project successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /projects/{projectId}/runs:
    get:
      tags: [Runs]
      operationId: listRuns
      summary: List runs for a project
      description: Returns runs ordered by createdAt desc.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/RunStatusFilter'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Ingestion]
      operationId: createRun
      summary: Create a run
      description: Creates a new run with status QUEUED.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRunResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/runs/{runId}:
    get:
      tags: [Runs]
      operationId: getRun
      summary: Get run details
      description: Returns a run that belongs to the resolved project.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetails'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Runs]
      operationId: deleteRun
      summary: Delete a run
      description: |
        Deletes a run and all associated test results.
        This operation cannot be undone.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RunId'
      responses:
        '204':
          description: No Content - Run successfully deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/runs/{runId}/results:
    get:
      tags: [Results]
      operationId: listRunResults
      summary: List results for a run
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResultListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/runs/{runId}/results/batch:
    post:
      tags: [Ingestion]
      operationId: batchIngestResults
      summary: Batch ingest results into a run
      description: |
        Upserts TestCase records (by projectId + externalId) and inserts TestResult rows for the run.
        Updates aggregated counters on the run (totalCount, passedCount, failedCount, skippedCount, errorCount).
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchResultsRequest'
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResultsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- Tests ----------
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authenticating requests.
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      description: Project slug or database id (implementation accepts both).
      schema:
        type: string
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Cursor:
      name: cursor
      in: query
      required: false
      description: Cursor pagination using the last seen run id.
      schema:
        type: string
    RunStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/RunStatus'
...
  responses:
    Unauthorized:
      description: Unauthorized (missing or invalid API key)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                statusCode: 401
                error: Unauthorized
                message: Authentication required
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                statusCode: 404
                error: Not Found
                message: Not found
    BadRequest:
      description: Bad request (validation error)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation error

    # ---------- Projects ----------
    Project:
      type: object
      required: [id, name, slug, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        createdAt:
          type: string
          format: date-time
      additionalProperties: false
    ProjectListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
      additionalProperties: false
    CreateProjectRequest:
      type: object
      required: [name, slug]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: string
          minLength: 1

    # ---------- Runs & Results ----------
    RunStatus:
      type: string
      enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELED]
    TestStatus:
      type: string
      enum: [PASSED, FAILED, SKIPPED, ERROR]
...
    RunListItem:
      type: object
      required:
        - id
        - createdAt
        - status
        - totalCount
        - passedCount
        - failedCount
        - skippedCount
        - errorCount
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/RunStatus'
        source:
          type: string
          nullable: true
        commitSha:
          type: string
          nullable: true
        branch:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          nullable: true
        totalCount:
          type: integer
        passedCount:
          type: integer
        failedCount:
          type: integer
        skippedCount:
          type: integer
        errorCount:
          type: integer
    RunListResponse:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RunListItem'
        nextCursor:
          type: string
          nullable: true
    RunDetails:
      type: object
      required:
        - id
        - projectId
        - status
        - createdAt
        - totalCount
        - passedCount
        - failedCount
        - skippedCount
        - errorCount
      properties:
        id:
          type: string
        projectId:
          type: string
        status:
          $ref: '#/components/schemas/RunStatus'
        createdAt:
          type: string
          format: date-time
        source:
          type: string
          nullable: true
        branch:
          type: string
          nullable: true
        commitSha:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          nullable: true
        totalCount:
          type: integer
        passedCount:
          type: integer
        failedCount:
          type: integer
        skippedCount:
          type: integer
        errorCount:
          type: integer
      additionalProperties: false
    CreateRunRequest:
      type: object
      additionalProperties: false
      properties:
        source:
          type: string
          example: manual
        commitSha:
          type: string
          example: deadbeef
        branch:
          type: string
          example: main
        env:
          type: object
          additionalProperties: true
          nullable: true
        meta:
          type: object
          additionalProperties: true
          nullable: true
    CreateRunResponse:
      type: object
      required: [id, createdAt, status, projectId]
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/RunStatus'
        projectId:
          type: string
      additionalProperties: false
    TestCaseRef:
      type: object
      required: [id, externalId, name, tags]
      properties:
        id:
          type: string
        externalId:
          type: string
        name:
          type: string
        suiteName:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      additionalProperties: false
...
    RunResultItem:
      type: object
      required: [id, status, createdAt, testCase]
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/TestStatus'
        durationMs:
          type: integer
          nullable: true
        message:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        testCase:
          $ref: '#/components/schemas/TestCaseRef'
      additionalProperties: false
    RunResultListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RunResultItem'
      additionalProperties: false
    BatchResultsRequest:
      type: object
      required: [results]
      properties:
        results:
          type: array
          minItems: 1
          items:
            type: object
            required: [externalId, name, status]
            properties:
              externalId:
                type: string
              name:
                type: string
              status:
                $ref: '#/components/schemas/TestStatus'
              durationMs:
                type: integer
                minimum: 0
              message:
                type: string
              stacktrace:
                type: string
              stdout:
                type: string
              stderr:
                type: string
              filePath:
                type: string
              suiteName:
                type: string
              tags:
                type: array
                items:
                  type: string
              meta:
                type: object
                additionalProperties: true
      additionalProperties: false
    BatchResultsResponse:
      type: object
      required: [inserted]
      properties:
        inserted:
          type: integer
          example: 2
      additionalProperties: false
...